refactor: Merge search and org reducers
          Rationale: search will give a different view into the org
          file. Hence it needs access to the same state.
