+++ src/reducers/org.js
@@ -1,6 +1,8 @@
 import { Map, List, fromJS } from 'immutable';
 import _ from 'lodash';
 
+import headline_filter_parser from '../lib/headline_filter_parser';
+
 import {
   parseOrg,
   parseTitleLine,
@@ -880,6 +882,19 @@ export const updateLogEntryTime = (state, action) => {
   );
 };
 
+export const setSearchFilter = (state, action) => {
+  const { searchFilter } = action;
+  try {
+    const filterExpr = headline_filter_parser.parse(searchFilter);
+    state = state.setIn(['search', 'searchFilterExpr'], filterExpr);
+  } catch (e) {
+    console.warn('Exception parsing headline filter: ' + e);
+  }
+  state = state.setIn(['search', 'searchFilter'], _.trim(searchFilter));
+
+  return state;
+};
+
 const setOrgFileErrorMessage = (state, action) => state.set('orgFileErrorMessage', action.message);
 
 export default (state = new Map(), action) => {
@@ -992,6 +1007,8 @@ export default (state = new Map(), action) => {
       return createLogEntryStart(state, action);
     case 'UPDATE_LOG_ENTRY_TIME':
       return updateLogEntryTime(state, action);
+    case 'SET_SEARCH_FILTER':
+      return setSearchFilter(state, action);
 
     default:
       return state;
